backend/requirements.txt
# Server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0

# Pipecat core + services
pipecat-ai[webrtc,deepgram,groq,silero]>=0.0.40

# Async utilities
aiohttp>=3.9.0